requirements.txt
spotipy
python-dotenv
pandas
seaborn